+++ CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to this project will be documented in this file.
 This project follows Semantic Versioning.
 
 ## Unreleased
+- Smooth active state to prevent animation flicker.
 
 ## 0.1.0 - 2026-01-24
 - Initial public release.